moved common files and add list of jobs link in header
