+++ static/js/common/ChangeStatus/index.js
@@ -1,0 +1,3 @@
+import ChangeStatus from './ChangeStatus';
+
+export default ChangeStatus;
