+++ static/js/common/ChangeStatus/switchButtonToStatus.js
