+++ static/js/common/switchColorToElement.js
@@ -2,7 +2,7 @@ import {
     STATUS_NEW, STATUS_TAKEN, STATUS_READY, STATUS_DELIVERED, STATUS_PAID, STATUS_CANCELED, STATUS_REOPENED,
     COLOR_NEW, COLOR_TAKEN, COLOR_READY, COLOR_DELIVERED, COLOR_PAID, COLOR_CANCELED, COLOR_REOPENED,
     BUTTON_NEW, BUTTON_TAKE, BUTTON_READY, BUTTON_DELIVER, BUTTON_PAY, BUTTON_CANCEL, BUTTON_REOPEN,
-} from '../../../../config';
+} from '../../../config';
 
 const switchColorToElement = (button) => {
     let currentColor = null;
